more outsourcing of utils
