+++ sbb_newspapers_org_image/utils/resize.py
@@ -1,0 +1,4 @@
+import cv2
+
+def resize_image(img_in, input_height, input_width):
+    return cv2.resize(img_in, (input_width, input_height), interpolation=cv2.INTER_NEAREST)
